requirements.txt
absl-py==2.1.0
-e git+ssh://git@forge-2.ircam.fr/acids/team/collaboration/acids-datasets.git@1bea66d4974bb758cf98a647f77e90e43d1d194b#egg=acids_datasets
anaconda-anon-usage @ file:///croot/anaconda-anon-usage_1710965072196/work
anyio==4.4.0
archspec @ file:///croot/archspec_1709217642129/work
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
audioread==3.0.1
Babel==2.15.0
-e git+ssh://git@forge-2.ircam.fr/acids/team/collaboration/acids-datasets.git@1bea66d4974bb758cf98a647f77e90e43d1d194b#egg=basic_pitch&subdirectory=acids_datasets/basic-pitch
beautifulsoup4==4.12.3
bleach==6.1.0
boltons @ file:///work/ci_py311/boltons_1677685195580/work
Brotli @ file:///croot/brotli-split_1714483155106/work
cachetools==5.3.3
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1718025014955/work/certifi
cffi @ file:///croot/cffi_1714483155441/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
cmake==3.25.0
comm==0.2.2
conda @ file:///home/conda/feedstock_root/build_artifacts/conda_1715631928155/work
conda-content-trust @ file:///croot/conda-content-trust_1714483159009/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1706733287605/work/src
conda-package-handling @ file:///croot/conda-package-handling_1714483155348/work
conda_package_streaming @ file:///croot/conda-package-streaming_1690987966409/work
contourpy==1.2.1
cryptography @ file:///croot/cryptography_1714660666131/work
cycler==0.12.1
Cython==3.0.10
DALI-dataset==1.1
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
distro @ file:///croot/distro_1714488253808/work
einops==0.8.0
encodec==0.1.1
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.15.4
flatbuffers==24.3.25
fonttools==4.53.0
fqdn==1.5.1
frozendict @ file:///home/conda/feedstock_root/build_artifacts/frozendict_1715092776266/work
fsspec==2024.6.0
future==1.0.0
gast==0.5.5
google-api-core==2.19.1
google-api-python-client==2.134.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
googleapis-common-protos==1.63.2
grpcio==1.64.1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.24.6
idna @ file:///croot/idna_1714398848350/work
ipykernel==6.29.4
ipython==8.25.0
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpatch @ file:///croot/jsonpatch_1714483231291/work
jsonpointer==2.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.1
jupyter_server_terminals==0.5.3
jupyterlab==4.2.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
keras==3.4.0
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
libmambapy @ file:///croot/mamba-split_1714483352891/work/libmambapy
librosa==0.10.2.post1
lit==15.0.7
llvmlite==0.43.0
lmdb==1.4.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
menuinst @ file:///croot/menuinst_1714510563922/work
mido==1.3.2
mir_eval==0.7
mistune==3.0.2
ml-dtypes==0.3.2
mpmath==1.3.0
msgpack==1.0.8
music2latent==0.1.6
mutagen==1.47.0
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nnAudio==0.3.1
notebook_shim==0.2.4
numba==0.60.0
numpy==1.23.5
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.7.0.84
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.535.161
nvidia-nccl-cu11==2.20.5
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.1.105
nvitop==1.3.2
oauthlib==3.2.2
opt-einsum==3.3.0
optree==0.11.0
overrides==7.7.0
packaging @ file:///croot/packaging_1710807400464/work
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
platformdirs @ file:///croot/platformdirs_1692205439124/work
pluggy @ file:///work/ci_py311/pluggy_1676822818071/work
pooch==1.8.2
pretty_midi==0.2.10
prometheus_client==0.20.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.3
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycosat @ file:///croot/pycosat_1714510623388/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pycryptodomex==3.20.0
Pygments==2.18.0
pyparsing==3.1.2
PySocks @ file:///work/ci_py311/pysocks_1676822712504/work
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
resampy==0.4.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
ruamel.yaml @ file:///work/ci_py311/ruamel.yaml_1676838772170/work
schedule==1.2.2
scikit-learn==1.5.0
scipy==1.14.0
Send2Trash==1.8.3
six==1.16.0
sniffio==1.3.1
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
stack-data==0.6.3
sympy==1.12.1
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.37.0
termcolor==2.4.0
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.3.0
torch==2.3.1
torchaudio==2.3.1
torcheval==0.0.7
torchinfo==1.8.0
torchvision==0.18.1
tornado==6.4.1
tqdm @ file:///croot/tqdm_1714567712644/work
traitlets==5.14.3
triton==2.3.1
truststore @ file:///croot/truststore_1695244293384/work
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
umap==0.1.1
uri-template==1.3.0
uritemplate==4.1.1
urllib3 @ file:///croot/urllib3_1707770551213/work
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.3
wrapt==1.14.1
yt-dlp==2024.7.2
zstandard @ file:///croot/zstandard_1714677652653/work
